extract it, like, way more
